+++ src/@types/dex-data.ts
@@ -1,0 +1,13 @@
+export interface DexData {
+  [key: number]: DexEntry;
+}
+
+export interface DexEntry {
+  seenAttr: bigint;
+  caughtAttr: bigint;
+  natureAttr: number;
+  seenCount: number;
+  caughtCount: number;
+  hatchedCount: number;
+  ivs: number[];
+}
