+++ src/data/abilities/ability-class.ts
@@ -2,7 +2,7 @@ import { AbilityId } from "#enums/ability-id";
 import type { AbAttrCondition } from "#app/@types/ability-types";
 import type { AbAttr } from "#app/data/abilities/ab-attrs/ab-attr";
 import i18next from "i18next";
-import type { Localizable } from "#app/interfaces/locales";
+import type { Localizable } from "#app/@types/locales";
 import type { Constructor } from "#app/utils/common";
 
 export class Ability implements Localizable {
