+++ src/data/pokemon-species.ts
@@ -1,4 +1,4 @@
-import type { Localizable } from "#app/interfaces/locales";
+import type { Localizable } from "#app/@types/locales";
 import { AbilityId } from "#enums/ability-id";
 import { PartyMemberStrength } from "#enums/party-member-strength";
 import { SpeciesId } from "#enums/species-id";
