+++ src/ui/pokemon-info-container.ts
@@ -6,7 +6,8 @@ import { getNatureName } from "../data/nature";
 import { PokemonType } from "#enums/pokemon-type";
 import type Pokemon from "../field/pokemon";
 import i18next from "i18next";
-import type { DexEntry, StarterDataEntry } from "../system/game-data";
+import type { StarterDataEntry } from "../system/game-data";
+import type { DexEntry } from "#app/@types/dex-data";
 import { DexAttr } from "../system/game-data";
 import { fixedInt, getShinyDescriptor } from "#app/utils/common";
 import ConfirmUiHandler from "./confirm-ui-handler";
